requirements.txt
Flask
pymongo
requests
pyyaml